requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
azure-core==1.30.1
azure-storage-blob==12.19.1
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
dnspython==2.6.1
email_validator==2.1.1
environs==9.5.0
essentia-tensorflow==2.1b6.dev1110
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastapi-login==1.10.0
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.25.0
isodate==0.6.1
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kapre==0.3.7
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.5
mkdocs==1.6.0 
ml-dtypes==0.3.2
moviepy==1.0.3
msgpack==1.0.8
music-tag==0.4.3
mutagen==1.47.0
mypy==1.9.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openl3==0.4.2
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.4.0
pooch==1.8.2
proglog==0.1.10
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.45
protobuf==3.20.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.18.0
PyJWT==2.8.0
pymilvus==2.4.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.4
requests==2.31.0
resampy==0.2.2
rich==13.7.1
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
spotipy==2.23.0
SQLAlchemy==2.0.29
sqlalchemy-stubs==0.4
stack-data==0.6.3
starlette==0.36.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tomli==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
trio==0.21.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
mkdocs
mkdocstrings
